--- src/main/java/ru/podgoretskaya/accounting/service/DayCalculation.java
@@ -1,12 +1,0 @@
-package ru.podgoretskaya.accounting.service;
-
-import ru.podgoretskaya.accounting.dto.PersonDTO;
-
-public interface DayCalculation {
-    long calculationWorkDay(PersonDTO p);
-
-    long daysOfSickDays(PersonDTO p);
-
-    long daysOfVacation(PersonDTO p);
-
-}
